complete customer

+++ server/src/modules/customers/infra/typeorm/repositories/CustomerRepository.ts
@@ -1,7 +1,8 @@
-import { getRepository, Repository } from 'typeorm';
+import { getRepository, Repository, Not } from 'typeorm';
 
 import ICreateCustomerDTO from '@modules/customers/dtos/ICreateCustomerDTO';
 import ICustomerRepository from '@modules/customers/repositories/ICustomerRepository'
+
 import Customer from '../entities/Customer';
 
 class CustomerRepository implements ICustomerRepository {
@@ -12,21 +13,31 @@ class CustomerRepository implements ICustomerRepository {
   }
 
   public async create(data: ICreateCustomerDTO): Promise<Customer> {
-    const customer = await this.ormRepository.create(data);
+    const customer = this.ormRepository.create(data);
 
     await this.ormRepository.save(customer);
 
     return customer;
   };
 
-  public async delete(id: string): Promise<void> {
-    await this.ormRepository.delete(id);
+  public async delete(user_id: string): Promise<void> {
+    const customer = await this.findById(user_id);
+
+    if (customer) {
+      await this.ormRepository.delete({ id: user_id });
+    }
   }
 
-  public async list(): Promise<Customer[]> {
+  public async findAllCustomers(): Promise<Customer[]> {
     const customers = await this.ormRepository.find();
 
     return customers;
+  }
+
+  public async findById(id: string): Promise<Customer | undefined> {
+    const customer = await this.ormRepository.findOne(id)
+
+    return customer;
   }
 
   public async findByCpf(cpf: number): Promise<Customer | undefined> {
@@ -39,7 +50,7 @@ class CustomerRepository implements ICustomerRepository {
     return customer;
   }
 
-  public async update(customer: Customer): Promise<Customer> {
+  public async save(customer: Customer): Promise<Customer> {
     const updatedCustomer = await this.ormRepository.save(customer);
 
     return updatedCustomer;
